refactor: Modify Ladder Machine apply feedback, query move to domain
